# go-fuzz: add SSE2 and AVX2 implementations of compareCoverBody