--- a/go-fuzz/compare_amd64.go
+++ b/go-fuzz/compare_amd64.go
@@ -4,7 +4,11 @@
 package main
 
 func compareCoverBody(base, cur []byte) bool {
-	return compareCoverBody1(&base[0], &cur[0])
+	if hasAVX2 {
+		return compareCoverBodyAVX2(&base[0], &cur[0])
+	}
+	return compareCoverBodySSE2(&base[0], &cur[0])
 }
 
-func compareCoverBody1(base, cur *byte) bool // in compare_amd64.s
+func compareCoverBodySSE2(base, cur *byte) bool // in compare_amd64.s
+func compareCoverBodyAVX2(base, cur *byte) bool // in compare_amd64.s