The SSE2 implementation uses only instructions that are assumed
available by the Go runtime. It can thus be used anywhere.

AVX2 support is checked for at runtime.

Benchmark Go implementation vs previous asm

name                     old time/op  new time/op  delta
CompareCoverBody/easy-8  6.79ns ± 9%  3.35ns ± 1%  -50.73%  (p=0.000 n=10+10)
CompareCoverBody/hard-8  18.6µs ± 5%   9.1µs ± 3%  -51.00%  (p=0.000 n=10+10)

Benchmark previous asm vs SSE2 asm

name                     old time/op  new time/op  delta
CompareCoverBody/easy-8  3.35ns ± 1%  5.56ns ± 5%  +66.01%  (p=0.000 n=10+10)
CompareCoverBody/hard-8  9.12µs ± 3%  1.79µs ± 2%  -80.38%  (p=0.000 n=10+9)

Benchmark SSE2 asm vs AVX2 asm

name                     old time/op  new time/op  delta
CompareCoverBody/easy-8  5.56ns ± 5%  5.35ns ± 2%   -3.82%  (p=0.000 n=10+9)
CompareCoverBody/hard-8  1.79µs ± 2%  1.05µs ± 2%  -41.36%  (p=0.000 n=9+10)

The regression in the easy benchmarks are almost entirely due
to the runtime check for AVX2 support.
It is also not particularly relevant to go-fuzz performance;
the hard case is both more typical and more important.