@josharian 2019-03-01
I’d like to also add an AVX512BW version, since compareCoverBody still manages to show up in profiles with AVX2, but my available hardware doesn’t support that instruction set.