@dvyukov 2019-03-02
One of the ideas for speeding it up that I had is moving the inner loop into the test binary somehow. Currently we have overhead of writing/reading pipe twice and switching between processes per test.  We can't move much into the test binary because we can't use lots of packages because they will need to be not instrumented then. But it should be possible to move the mutation logic and cover comparison into the binary. And maybe it's actually fine if some deps will be instrumented, they will be somewhat slower, but if we reset cover before calling into the Fuzz function (which we do anyways) and there are no background goroutines then it should be fine.