@josharian 2019-03-02
I’ve had that thought as well. But I’ve experimented with some fairly unusual mutation strategies; worrying about the impact of deps there would be unfortunate. So I’m personally not particularly inclined to do that, at least not right now.